Extract upgradeExpression as a visitor, set up upgradeExpressions as the principal driver of upgrading
