+++ src/test/kotlin/org/openrewrite/terraform/terraform012/UseUpdatedConnectionBlockTest.kt
@@ -27,7 +27,7 @@ import org.openrewrite.hcl.HclRecipeTest
 @Suppress("RemoveCurlyBracesFromTemplate")
 class UseUpdatedConnectionBlockTest : HclRecipeTest {
     override val recipe: Recipe
-        get() = UseFirstClassExpressions()
+        get() = UpgradeExpressions()
 
     @Test
     @Disabled
